+++ src/main/java/gov/nist/oar/distrib/datapackage/DataPackager.java
@@ -10,14 +10,14 @@
  * that they have been modified.
  * @author: Deoyani Nandrekar-Heinis
  */
-package gov.nist.oar.distrib;
+package gov.nist.oar.distrib.datapackage;
 
 import java.io.IOException;
 import java.io.OutputStream;
 import java.net.MalformedURLException;
 import java.util.zip.ZipOutputStream;
 
-import gov.nist.oar.distrib.datapackage.BundleDownloadPlan;
+import gov.nist.oar.distrib.DistributionException;
 
 /**
  * DataPackager interface declares the methods to be used when data download
